Added composite mass storage/CDC example
This allows for the device usb on the board to act both as a virtual
serial port for bidirectional comms and also a mass storage device. It
needs the 'usb_dev_chidcdc.inf' driver to run which is in
StellarisWare\windows_drivers.

It's basically two examples hobbled together but it was incredibly
fiddly to get working so thought I'd commit a copy now before I break
something and can't get it to work any more. Code is very messy at the
moment and I'm pretty sure there's tons of stuff in there that we don't
actually need. Hoping that I won't break anything by committing it..